requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
line-bot-sdk==3.13.0